Conflicts:
	src/cc1/Makefile.dep
	src/cc1/cc1.h
	src/cc1/decl_init.c
	src/cc1/fold.c
	src/cc1/gen_asm.c
	src/cc1/out/impl.c
	src/cc1/out/impl.h
	src/cc1/str.c

+++ src/cc1/str.h
@@ -1,7 +1,7 @@
 #ifndef STR_H
 #define STR_H
 
-void escape_string(char *str, int *len);
+void escape_string(char *str, size_t *len);
 
 char *str_add_escape(const char *s, const size_t len);
 int literal_print(FILE *f, const char *s, size_t len);
